+++ CHANGELOG.md
@@ -7,12 +7,51 @@ that can be found in the LICENSE file. -->
 > [!IMPORTANT]  
 > See the [Migration Guide](guides/migration_guide.md) for the details of breaking changes between versions.
 
+## Unreleased
+
+*None.*
+
+## 9.2.1
+
+- Improve changes when limited on iOS.
+- Use `LocallyAvailableBuilder` in the grid to provide better user awareness.
+
+## 9.2.0
+
+### New features
+
+- Introduce `AssetsChangeCallback` and `AssetsChangeRefreshPredicate`
+  to help users act according to assets changes.
+- Add `shouldAutoplayPreview` in the config.
+
+### Fixes
+
+- Raise detailed negative range error.
+- Fix viewer confirm button predication.
+- Enlarge GIF gradients.
+- Fix potential paths assets count unexpected merging behaviors.
+
+## 9.1.0
+
+### Improvements
+
+- Support limited permission displays on Android.
+- Improves the limited overlay padding on Android.
+- Adds permission request lock for the picker state.
+- Speeding up by splitting asset loading into separate steps.
+- Speeding up using `AdvancedCustomFilter` rather than `FilterOptionGroup` by default.
+
+### Fixes
+
+- Fix reverted index when previewing assets on Android.
+- Requests with the correct options with the picker.
+
 ## 9.0.4
 
 ### Fixes
 
 - Fix the app bar of the viewer that is not animating.
-- Fix loading when no assets in the path.
+- Fix loading when no assets are in the path.
 - Reset the has more to load flag between path switching.
 
 ## 9.0.3
@@ -31,7 +70,7 @@ that can be found in the LICENSE file. -->
 
 ### Fixes
 
-- Fix current asset in the picker viewer.
+- Fix the current asset in the picker viewer.
 
 ### Improvements